requirements.txt
supabase
python-dotenv
google-generativeai
